refactor: split command handlers out of main.rs into focused modules (#11)
Move 35 Tauri command handlers from main.rs into organized modules:
- commands/sessions.rs — session/PTY lifecycle, git helpers
- commands/worktrees.rs — worktree CRUD, branches, git init
- commands/projects.rs — project CRUD, notes
- commands/docs.rs — file read/write, doc listing
- commands/setup.rs — setup status, nono profiles
- commands/settings.rs — get/update settings
- commands/misc.rs — logging, editor, quit

Also extracts AppState into state.rs with pub(crate) fields.

main.rs drops from 815 to 160 lines — now just wiring:
mod declarations, builder, plugins, state init, generate_handler!, and
setup/event handlers.

No logic changes, no frontend changes. Pure mechanical move.

Closes #11
