+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod docs;
+pub(crate) mod misc;
+pub(crate) mod projects;
+pub(crate) mod sessions;
+pub(crate) mod settings;
+pub(crate) mod setup;
+pub(crate) mod worktrees;
